feat: member 기능 구현

+++ src/main/java/com/fivemybab/ittabab/member/command/repository/MemberCommandRepository.java
@@ -2,6 +2,6 @@ package com.fivemybab.ittabab.member.command.repository;
 
 import com.fivemybab.ittabab.member.command.entity.MemberInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberCommandRepository extends JpaRepository<MemberInfo, Integer> {
+// CUD 작업 진행
+public interface MemberCommandRepository extends JpaRepository<MemberInfo,Long> {
 }
